+++ src/dt4acc_lib/bl/translator_service.py
@@ -2,11 +2,14 @@ import functools
 import logging
 from typing import Union
 
+from .remap_and_convert import RemapIdentifiersAndConvert
 from .tune_translator import TuneConversion
 from dt4acc_lib.bl.identity_conversion import identity_conversion
 from dt4acc_lib.bl.unit_conversion import (
     LinearUnitConversion,
     EnergyDependentLinearUnitConversion,
+    MultiplierScaledByEnergyUnitConversion,
+    UnitConversionAroundReferenceValue,
 )
 from dt4acc_lib.interfaces.utils.state_conversion import StateConversion
 from dt4acc_lib.interfaces.utils.translator_service import TranslatorServiceBase
@@ -16,6 +19,9 @@ from dt4acc_lib.model.utils.translator_manager_lookup_table import (
     PolynomCoefficients,
     TuneConversionCoefficients,
     IdentityMapper,
+    MultiplyerScaledByEnergy,
+    NeedsAReference,
+    RemapIdentifiersAndConvertDM
 )
 
 
@@ -36,6 +42,8 @@ class TranslatorService(TranslatorServiceBase):
     def __init__(self, *, lut: TranslatorLookupTable, brho: float):
         self.lut = lut
         self.brho = brho
+        self.reference_cache = None
+        self.to_needs_reference_cache = []
 
     def __str__(self):
         return f"{self.__class__.__name__}(brho=self.brho, lut with {len(self.lut.lut)} entries)"
@@ -48,6 +56,18 @@ class TranslatorService(TranslatorServiceBase):
             f", lut={repr(self.lut)}"
         )
 
+    def register_reference_cache(self, cache):
+        self.reference_cache = cache
+        for to in self.to_needs_reference_cache:
+            to.register_reference_cache(self.reference_cache)
+
+    def needs_reference_cache(self, to):
+        assert callable(to.register_reference_cache)
+        if to not in self.to_needs_reference_cache:
+            self.to_needs_reference_cache.append(to)
+        if self.reference_cache:
+            to.register_reference_cache(self.reference_cache)
+
     @functools.lru_cache(maxsize=None)
     def get(self, id_: ConversionID) -> Union[StateConversion, None]:
         to = self._lookup_config(id_)
@@ -57,6 +77,14 @@ class TranslatorService(TranslatorServiceBase):
         # Should one use a factory down here ?
         # Then it could be handled over and configured by the user
         r = create_translation_object_for_data_model(to, self.brho)
+        # Todo: need to improve which objects need a reference
+        #       furthermore, perhaps it should only be a week
+        #       reference here, so it gets out of this list
+        #       when not needed anymore else ....
+        if isinstance(to, NeedsAReference) or isinstance(r, UnitConversionAroundReferenceValue):
+            assert isinstance(to, NeedsAReference)
+            assert isinstance(r, UnitConversionAroundReferenceValue)
+            self.needs_reference_cache(r)
         return r
 
     def _lookup_config(self, id_: ConversionID):
@@ -91,7 +119,10 @@ class TranslatorService(TranslatorServiceBase):
 
 
 def create_translation_object_for_data_model(
-    data_model: Union[PolynomCoefficients | TuneConversionCoefficients], brho: float
+    data_model: Union[
+        PolynomCoefficients | TuneConversionCoefficients, MultiplyerScaledByEnergy, NeedsAReference
+    ],
+    brho: float,
 ):
     if isinstance(data_model, IdentityMapper):
         return identity_conversion
@@ -99,11 +130,24 @@ def create_translation_object_for_data_model(
         return create_translation_object_for_polynom_coefficients(data_model, brho)
     elif isinstance(data_model, TuneConversionCoefficients):
         return create_translation_object_for_tune(data_model)
+    elif isinstance(data_model, MultiplyerScaledByEnergy):
+        return create_translation_object_for_multiplyer_scaled_by_energy(
+            data_model, brho
+        )
+    elif isinstance(data_model, NeedsAReference):
+        return create_translation_object_for_needs_a_reference(data_model, brho)
+    elif isinstance(data_model, RemapIdentifiersAndConvertDM):
+        return create_translation_object_for_remap_and_convert(data_model, brho)
     else:
         # Todo: fix the error that is raised
         raise AssertionError(
             f"Don't know how to instantiate data_model of {type(data_model)} for {data_model}"
         )
+
+def create_translation_object_for_remap_and_convert(data_model: RemapIdentifiersAndConvertDM, brho: float):
+    conv = create_translation_object_for_polynom_coefficients(data_model.conversion, brho)
+    r = RemapIdentifiersAndConvert(data_model.name_mapping, conv)
+    return r
 
 
 def create_translation_object_for_tune(
@@ -131,5 +175,29 @@ def create_translation_object_for_polynom_coefficients(
     else:
         return LinearUnitConversion(slope=slope, intercept=intercept)
 
+
+def create_translation_object_for_multiplyer_scaled_by_energy(
+    data_model: MultiplyerScaledByEnergy, brho: float
+) -> MultiplierScaledByEnergyUnitConversion:
+    return MultiplierScaledByEnergyUnitConversion(conv_data=data_model, brho=brho)
+
+
+def create_translation_object_for_needs_a_reference(
+        data_model: NeedsAReference, brho: float
+) -> UnitConversionAroundReferenceValue:
+    """
+    Todo:
+        should it be able to look up the conversion for the given
+        data model.translation_object?
+        It could exist already
+    """
+    to = create_translation_object_for_data_model(data_model.translation_object, brho)
+    return UnitConversionAroundReferenceValue(
+        sub_obj=to,
+        # forward_rcmd=data_model.forward_read_command,
+        design_view_rcmd=data_model.design_view_read_commnd,
+        # Todo: should be a specific read command
+        device_view_rcmd=data_model.device_view_read_command,
+    )
 
 __all__ = ["TranslatorService"]