+++ src/dt4acc_lib/bl/unit_conversion.py
@@ -3,10 +3,31 @@ import logging
 import numpy as np
 from typing import Sequence
 from dt4acc_lib.interfaces.utils.state_conversion import StateConversion
-from dt4acc_lib.model.utils.identifiers import CurvePoint
+from dt4acc_lib.model.utils.translator_manager_lookup_table import (
+    CurvePoint,
+    MultiplyerScaledByEnergy,
+)
 from scipy.interpolate import interp1d
 
+from dt4acc_lib.model.utils.command import ReadCommand
+
 logger = logging.getLogger("dt4acc_lib")
+
+
+def calculate_brho(energy: float, rest_mass: float = 511e3) -> float:
+    """
+    Todo:
+        Energy is assumed here to be in eV not GeV
+    """
+    energy_without_rest_mass = np.sqrt((energy + rest_mass) ** 2 - rest_mass ** 2)
+    brho = energy_without_rest_mass / speed_of_light
+    return brho
+
+
+def calculate_brho_inverse(brho: float, rest_mass: float = 511e3) -> float:
+    energy_without_rest_mass = brho * speed_of_light
+    energy = np.sqrt((energy_without_rest_mass) ** 2 - rest_mass ** 2) - rest_mass
+    return energy
 
 
 class EnergyDependentLinearUnitConversion(StateConversion):
@@ -72,7 +93,6 @@ class LinearUnitConversion(StateConversion):
         return (state - self.intercept) / self.slope
 
 
-
 class EnergyIndependentCurveUnitConversion(StateConversion):
     """
     Interpolate a curve (independent -> dependent) using scipy.interpolate.interp1d
@@ -93,13 +113,14 @@ class EnergyIndependentCurveUnitConversion(StateConversion):
         # todo: see how it was named at max iv
         #       iimplement it in a filter delegating implementation to this object
         flip_dep_sign: False,
-        length : float
+        length: float
     ):
         # forward interpolator: will raise if x out of bounds
         # TODO: clean up it is a mess at the moment
+        # Todo: check that these conversions are correct
         self._fwd = interp1d(
-            [t.indep for t in fwd_points],
             [t.dep for t in fwd_points],
+            [t.indep for t in fwd_points],
             kind="linear",
             # Todo: change later to true ... or make user configurable
             bounds_error=False,
@@ -125,17 +146,115 @@ class EnergyIndependentCurveUnitConversion(StateConversion):
         y = float(self._fwd(x))  # interp1d returns an array-like
         if self.flip_dep_sign:
             y = -y
-        return  y * self.brho
+        return y * self.brho
 
     def inverse(self, state: float) -> float:
         # logger.info("%s.inverse: brho %s points %d state %s", self.__class__.__name__, self.brho, len(self._indep), state)
         if self.brho == 0:
             raise ValueError("brho must be non-zero for inversion")
         target = float(state) / self.brho
-        y = float(self._bwd(target)) # interp1d returns an array-like
+        y = float(self._bwd(target))  # interp1d returns an array-like
         assert np.isfinite(y), "failed to inverse {state=} ({brho=}, {target=})"
         if self.flip_dep_sign:
             y = y * -1
         return y / self.length
 
-__all__ = ["EnergyDependentLinearUnitConversion", "LinearUnitConversion", "EnergyIndependentCurveUnitConversion"]
+
+class MultiplierScaledByEnergyUnitConversion(StateConversion):
+    def __init__(self, *, conv_data: MultiplyerScaledByEnergy, brho: float):
+        """Todo: need actual brho or energy"""
+        self.brho = float(brho)
+        self.conv_data = conv_data
+
+        self.fwd_interp = interp1d(
+            # from energy to scaled current
+            [t.indep for t in conv_data.scale_by_energy],
+            [t.dep for t in conv_data.scale_by_energy],
+            kind="linear",
+            # Todo: change later to true ... or make user configurable
+            bounds_error=False,
+            fill_value="extrapolate"
+        )
+        self.bwd_interp = interp1d(
+            # from scaled current to energy
+            [t.dep for t in conv_data.scale_by_energy],
+            [t.indep for t in conv_data.scale_by_energy],
+            kind="linear",
+            # Todo: change later to true ... or make user configurable
+            bounds_error=False,
+            fill_value="extrapolate"
+        )
+
+    def forward(self, state: float) -> float:
+        """corresponds to k2amp"""
+        ref_field = state * self.brho
+        t_brho = ref_field / self.conv_data.reference_multiplyer
+        energy = calculate_brho_inverse(t_brho)
+        scale = self.fwd_interp(energy)
+        r = self.conv_data.range.interpolate(scale)
+        return r
+
+    def inverse(self, state: float) -> float:
+        """corresponds to amp2k"""
+        lambda_ = self.conv_data.range.scale(state)
+        corresponding_energy = self.bwd_interp(lambda_)
+        ref_k = self.conv_data.reference_multiplyer
+        ref_brho = calculate_brho(self.conv_data.reference_energy)
+        # Todo: fix data to be in eV
+        t_brho = calculate_brho(corresponding_energy)
+        correction = t_brho / ref_brho
+        return ref_k * correction
+
+
+class UnitConversionAroundReferenceValue(StateConversion):
+    """
+    Todo:
+        can we reuse state cache ...
+    """
+    def __init__(self, *, sub_obj: StateConversion, design_view_rcmd: ReadCommand, device_view_rcmd: ReadCommand, reference_cache = None):
+        """
+        Todo:
+            type reference_cache as StateCache
+        """
+        self.sub_obj = sub_obj
+        self.design_view_rcmd = design_view_rcmd
+        self.device_view_rcmd = device_view_rcmd
+        self.reference_cache = reference_cache
+
+    def register_reference_cache(self, cache):
+        assert callable(cache.get)
+        self.reference_cache = cache
+
+    def forward(self, state: float) -> float:
+        dval = float(state)
+        rev_val = self.get_reference_value(self.design_view_rcmd)
+        total_val = dval + rev_val
+        tmp = self.sub_obj.forward(total_val)
+        rev_tgt = self.get_reference_value(self.device_view_rcmd)
+        return tmp - rev_tgt
+
+    def inverse(self, state: float) -> float:
+        dval = float(state)
+        rev_val = self.get_reference_value(self.device_view_rcmd)
+        total_val = dval + rev_val
+        tmp = self.sub_obj.inverse(total_val)
+        rev_tgt = self.get_reference_value(self.design_view_rcmd)
+        return tmp - rev_tgt
+
+    def get_reference_value(self, rcmd: ReadCommand) -> float:
+        if not self.reference_cache:
+            raise ValueError("reference_cache must be registered at this stage. Did you register it with translation service?")
+        assert rcmd is not None
+        rev_val = self.reference_cache.get(rcmd)
+        if rev_val is None:
+            raise ValueError(f"reference_cache {self.reference_cache} must provide reference data for {rcmd} but returned None. Did you populate the cache with an appropriate value")
+        return rev_val
+
+__all__ = [
+    "EnergyDependentLinearUnitConversion",
+    "LinearUnitConversion",
+    "EnergyIndependentCurveUnitConversion",
+    "MultiplierScaledByEnergyUnitConversion",
+    "calculate_brho_inverse",
+    "calculate_brho",
+]