+++ Changelog.md
@@ -6,3 +6,11 @@
 * version 0.4.3 simulation engine: 
   fix that reinit method awaits method reset,
   so reset is only working properly since this version
+
+## for version 2
+### for version 2.0
+
+* [FIX] translation object EnergyIndependendtCurveUnitConversion
+  forward interpolation (lattice -> device): constructs forward interpolation from dep / indep points 
+  now it is the reverse of bwd / inverse
+